+++ virtcontainers/api.go
@@ -613,6 +613,7 @@ func statusContainer(sandbox *Sandbox, containerID string) (ContainerStatus, err
 				PID:         container.process.Pid,
 				StartTime:   container.process.StartTime,
 				RootFs:      container.config.RootFs.Target,
+				Spec:        container.GetOCISpec(),
 				Annotations: container.config.Annotations,
 			}, nil
 		}
