+++ virtcontainers/sandbox_test.go
@@ -649,6 +649,7 @@ func TestContainerStateSetFstype(t *testing.T) {
 		{
 			ID:          "100",
 			Annotations: containerAnnotations,
+			Spec:        newEmptySpec(),
 		},
 	}
 
@@ -1517,44 +1518,24 @@ func TestSandboxExperimentalFeature(t *testing.T) {
 	assert.True(t, sconfig.valid())
 }
 
-/*
-func TestSandbox_joinSandboxCgroup(t *testing.T) {
-
-	mockValidCgroup := &Sandbox{}
-	mockValidCgroup.state.CgroupPath = "/my/cgroup"
-
-	tests := []struct {
-		name    string
-		s       *Sandbox
-		wantErr bool
-	}{
-		{"New Config", &Sandbox{}, false},
-		{"Mock cgroup path", mockValidCgroup, false},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := tt.s.joinSandboxCgroup(); (err != nil) != tt.wantErr {
-				t.Errorf("Sandbox.joinSandboxCgroup() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-*/
-
 func TestSandbox_SetupSandboxCgroup(t *testing.T) {
 	sandboxContainer := ContainerConfig{}
 	sandboxContainer.Annotations = make(map[string]string)
 	sandboxContainer.Annotations[annotations.ContainerTypeKey] = string(PodSandbox)
 
-	emptyJSONLinux := ContainerConfig{}
+	emptyJSONLinux := ContainerConfig{
+		Spec: newEmptySpec(),
+	}
 	emptyJSONLinux.Annotations = make(map[string]string)
 	emptyJSONLinux.Annotations[annotations.ContainerTypeKey] = string(PodSandbox)
-	emptyJSONLinux.Annotations[annotations.ConfigJSONKey] = "{}"
 
-	successfulContainer := ContainerConfig{}
+	cloneSpec1 := newEmptySpec()
+	cloneSpec1.Linux.CgroupsPath = "/myRuntime/myContainer"
+	successfulContainer := ContainerConfig{
+		Spec: cloneSpec1,
+	}
 	successfulContainer.Annotations = make(map[string]string)
 	successfulContainer.Annotations[annotations.ContainerTypeKey] = string(PodSandbox)
-	successfulContainer.Annotations[annotations.ConfigJSONKey] = "{\"linux\": { \"cgroupsPath\": \"/myRuntime/myContainer\" }}"
 
 	tests := []struct {
 		name    string
